Added permissions column for agents, created different levels of roles and handled authorization for roles/permissions in endpoints

+++ triage_app/routes/group.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from .. import schemas
 from sqlalchemy.orm import Session
 from ..dependencies import get_db
-from ..crud import create_group, delete_group, update_group, decode_token, get_group_by_filter, get_groups
+from ..crud import create_group, delete_group, update_group, decode_token, get_group_by_filter, get_groups, get_permission
 from fastapi.responses import JSONResponse
 
 
@@ -10,6 +10,8 @@ router = APIRouter(prefix='/group')
 
 @router.post("/create", response_model=schemas.Group)
 def group_create(group: schemas.GroupCreate, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='org.create'):
+        raise HTTPException(status_code=403, detail=f"Access denied: You do not have permission to access this resource")
     return create_group(db=db, group=group)
 
 
@@ -26,6 +28,8 @@ def get_all_groups(db: Session = Depends(get_db), agent_data: schemas.TokenData 
 
 @router.put("/put/{group_id}", response_model=schemas.Group)
 def group_update(group_id: int, updates: schemas.GroupUpdate, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='org.edit'):
+        raise HTTPException(status_code=403, detail=f"Access denied: You do not have permission to access this resource")
     group = update_group(db, group_id, updates)
     if not group:
         raise HTTPException(status_code=400, detail=f'Group with id {group_id} not found')
@@ -34,6 +38,8 @@ def group_update(group_id: int, updates: schemas.GroupUpdate, db: Session = Depe
 
 @router.delete("/delete/{group_id}")
 def group_delete(group_id: int, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='org.delete'):
+        raise HTTPException(status_code=403, detail=f"Access denied: You do not have permission to access this resource")
     status = delete_group(db, group_id)
     if not status:
         raise HTTPException(status_code=400, detail=f'Group with id {group_id} not found')
