+++ triage_app/routes/user.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from .. import schemas
 from sqlalchemy.orm import Session
 from ..dependencies import get_db
-from ..crud import create_user, delete_user, update_user, decode_token, get_user_by_filter, get_users
+from ..crud import create_user, delete_user, update_user, decode_token, get_user_by_filter, get_users, get_permission
 from fastapi.responses import JSONResponse
 
 
@@ -10,11 +10,15 @@ router = APIRouter(prefix='/user')
 
 @router.post("/create", response_model=schemas.User)
 def user_create(user: schemas.UserCreate, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='user.create'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")
     return create_user(db=db, user=user)
 
 
 @router.get("/id/{user_id}", response_model=schemas.User)
 def get_user_by_id(user_id: int, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='user.manage'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")
     user = get_user_by_filter(db, filter={'user_id': user_id})
     if not user:
         raise HTTPException(status_code=400, detail=f'No user found with id {user_id}')
@@ -22,18 +26,23 @@ def get_user_by_id(user_id: int, db: Session = Depends(get_db), agent_data: sche
 
 @router.get("/get", response_model=list[schemas.User])
 def get_all_users(db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='user.dir'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")
     return get_users(db)
 
 @router.put("/put/{user_id}", response_model=schemas.User)
 def user_update(user_id: int, updates: schemas.UserUpdate, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='user.edit'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")
     user = update_user(db, user_id, updates)
     if not user:
         raise HTTPException(status_code=400, detail=f'User with id {user_id} not found')
-    
     return user
 
 @router.delete("/delete/{user_id}")
 def user_delete(user_id: int, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_permission(db, agent_id=agent_data.agent_id, permission='user.delete'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")
     status = delete_user(db, user_id)
     if not status:
         raise HTTPException(status_code=400, detail=f'User with id {user_id} not found')
